Fixed stuff
First of all we made the folders to modules and made the approp. changes to
the files. Further we moved the config.py to a different place because it
should not have been in the top-level folder in the first place.

We also added a README whic hdescribes how to setup and test the project/