Update main.yml
This should fix our problem. Git hub uses keys to check if a cache was invalidated, see here for details.

In a nutshell, I used the wrong file in the hash key, thus when our req. file changed git hub did not care because I did not check for those changes 
https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows#matching-a-cache-key